Recover the localization code in the getStorageData function
Recover the code that allows the display, in the getStorageData function, of the word Available in the different languages included in the app.
This code existed a few weeks ago but now I don't see it anymore.
If it has been intentionally removed, I ask you to re-evaluate its inclusion.